Add ProjectDeleteForm for project deletion confirmation
Introduces ProjectDeleteForm to require ADMIN (only) to confirm the project title and acknowledge the irreversible nature of deletion. This enhances safety and prevents accidental deletion of projects and related records. Note that this is for CE and is probably more CE oriented than it should be but the concept is that any "PROJECT" can be deleted entirely.