logic for action list revision implemented
It's important to note here that the current system allows for multiple versions of the same document, with provenance respected. But it might be the case where the document is hosted externally for AB to comment on a single instance hosted. however, this would violate the single host, single source of truth principle for this system. Will speak to stakeholders to evaluate and determine their decision.